ik heb de create melding ietsjes verbeterd, ik heb geprobeerd de naam te laten zien en de reactie. Het zou allemaal horen te werken. maar iets werkt niet lekker mee

+++ Website/Look/Models/Reactie.cs
@@ -10,7 +10,7 @@ namespace Look.Models
     public class Reactie
     {
         [Key]
-        public int ReactieId {get; set;}
+        public long ReactieId {get; set;}
         public Gebruiker GeplaatstDoor {get; set;}
         public DateTime GeplaatstOp {get; set;}
         public string Bericht {get; set;}
